+++ src/algorithm/monadic/multivector.rs
@@ -45,17 +45,17 @@ impl Value {
             (Some(2), None | Some(2), Some(SubSide::Left)) | (Some(2), Some(2), None) => {
                 Array::new(
                     new_shape,
-                    arr.data.chunks_exact(2).map(|n| Complex::new(n[0], n[1])),
+                    (arr.data.as_chunks::<2>().0.iter()).map(|n| Complex::new(n[0], n[1])),
                 )
                 .into()
             }
             #[cfg(not(feature = "ga"))]
             (Some(4), Some(2), None)
-                if arr.data.chunks_exact(4).all(|w| w[1] == 0.0 && w[2] == 0.0) =>
+                if (arr.data.as_chunks::<4>().0).all(|w| w[1] == 0.0 && w[2] == 0.0) =>
             {
                 Array::new(
                     new_shape,
-                    arr.data.chunks_exact(4).map(|n| Complex::new(n[0], n[3])),
+                    (arr.data.as_chunks::<4>().0).map(|n| Complex::new(n[0], n[3])),
                 )
                 .into()
             }