+++ src/lsp.rs
@@ -1196,24 +1196,22 @@ mod server {
                             underable,
                             pure,
                             ..
-                        } => {
-                            if pure || invertible || underable {
-                                value.push_str("\n\n");
+                        } if (pure || invertible || underable) => {
+                            value.push_str("\n\n");
+                            if pure {
+                                value.push_str("pure");
+                            }
+                            if invertible {
                                 if pure {
-                                    value.push_str("pure");
+                                    value.push_str(" | ");
                                 }
-                                if invertible {
-                                    if pure {
-                                        value.push_str(" | ");
-                                    }
-                                    value.push_str("[`° un`](https://uiua.org/docs/un)");
+                                value.push_str("[`° un`](https://uiua.org/docs/un)");
+                            }
+                            if underable {
+                                if pure || invertible {
+                                    value.push_str(" | ");
                                 }
-                                if underable {
-                                    if pure || invertible {
-                                        value.push_str(" | ");
-                                    }
-                                    value.push_str("[`⍜ under`](https://uiua.org/docs/under)");
-                                }
+                                value.push_str("[`⍜ under`](https://uiua.org/docs/under)");
                             }
                         }
                         _ => {}
@@ -1338,7 +1336,7 @@ mod server {
                         ..Default::default()
                     }),
                     sort_text: text.split('.').next_back().map(Into::into),
-                    filter_text: text.split('.').next_back().map(Into::into),
+                    // filter_text: text.split('.').next_back().map(Into::into),
                     documentation: binding.meta.comment.as_ref().map(|c| {
                         Documentation::MarkupContent(MarkupContent {
                             kind: MarkupKind::Markdown,